model/pyproject.toml
[tool.poetry]
name = "ai4mde-studio"
version = "1.0.0"
description = "AI-assisted Model Driven Engineering"
authors = ["Max Boone <contact@maxboone.com>"]
license = "GPLv3"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
django = "5.0.1"
psycopg2-binary = "^2.9.9"
django-ninja = "^1.1.0"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
ruff = "^0.1.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
